pyproject.toml
[project]
name = "sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bs4>=0.0.2",
    "docker>=7.1.0",
    "fastapi>=0.124.4",
    "lxml>=6.0.2",
    "pathlib>=1.0.1",
    "psycopg-binary>=3.2.13",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
]
